feat: реализован функционал RequestService
дописан код работы с запросами, исправлены ошибки по результатам тестов.

+++ main-svc/src/main/java/ru/practicum/ewm/mainsvc/event/dto/EventDto.java
@@ -3,5 +3,5 @@ package ru.practicum.ewm.mainsvc.event.dto;
 public interface EventDto {
     Long getId();
 
-    void setConfirmedRequests(Long count);
+    void setConfirmedRequests(Integer count);
 }
