+++ main-svc/src/main/java/ru/practicum/ewm/mainsvc/request/RequestCountView.java
@@ -3,5 +3,5 @@ package ru.practicum.ewm.mainsvc.request;
 public interface RequestCountView {
     Long getEventId();
 
-    Long getRequestCount();
+    Integer getRequestCount();
 }
